TST: be a bit more relaxed about test_alarms speed
